services/apigateway-py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
redis==5.2.0
pydantic==2.10.0
python-dotenv==1.0.0
pyyaml==6.0.2
httpx==0.27.0